Opacity & Skinning
Initially I was going to leave skinning until the end until I realized you were not able to use opacity without skinning (or adding images to sprites). Right now it's not running great locally (I'll need to figure that out) but works well on open processing.

I made little pixel animations for my setup and really like it! only problem is the placement changes (it still registers the transparent part of the transparent backgrounds). May need to throw that in photoshop.

I'm still working out if I like the current opacity, I'm going to try and change the image (as per your suggestion - thank you for that) with sprite information when it overlaps.

Also still need to add the timer and start and end game screen.